Fixed compilation under clang

+++ gripper_controllers/include/gripper_controllers/hardware_interface_adapter.hpp
@@ -170,7 +170,7 @@ public:
     // Time since the last call to update
     const auto period = last_update_time_ > std::chrono::steady_clock::time_point() ? std::chrono::steady_clock::now() - last_update_time_ : std::chrono::nanoseconds::zero();
     // Update PIDs
-    double raw_command = pid_->computeCommand(error_position, error_velocity, period.count());
+    double raw_command = pid_->computeCommand(error_position, error_velocity, (uint64_t)period.count());
     double command = std::min<double>(fabs(max_allowed_effort), std::max<double>(-fabs(max_allowed_effort), raw_command));
     //RCLCPP_INFO_STREAM(rclcpp::get_logger("gazebo_ros2_control"), joint_handle_->get().get_prefix_name() << ": raw_command: " << raw_command << ", command: " << command << ", max_allowed_effort: " << max_allowed_effort << ", error_position: " << error_position << ", error_velocity: " << error_velocity);
     joint_handle_->get().set_value(command);
